% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{build}
\alias{build}
\title{Build Final S3 Objects}
\usage{
build(x, ...)
}
\arguments{
\item{x}{The prepared data object to process.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A "built" object, typically a container for lists of final S3 objects.
}
\description{
A generic S3 function that dispatches to a specific method to
  take prepared data and assemble the final, render-ready S3 objects
  (e.g., `sem_node`, `sem_edge`).
}
\keyword{internal}
