% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{configure_plot}
\alias{configure_plot}
\title{Configure a Plot}
\usage{
configure_plot(x, ...)
}
\arguments{
\item{x}{The analyzed object to configure.}

\item{...}{User arguments for configuration.}
}
\value{
A final configuration object.
}
\description{
A generic S3 function that dispatches to a specific method to
  take analysis results and user arguments, and assembles the final
  configuration object (e.g., a recipe).
}
\keyword{internal}
