% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-preparation.R
\name{.apply_node_styling}
\alias{.apply_node_styling}
\title{Apply Node Styling from Recipe}
\usage{
.apply_node_styling(nodes_data, recipe)
}
\arguments{
\item{nodes_data}{The node data table.}

\item{recipe}{The final, complete recipe object.}
}
\value{
The `nodes_data` table with added columns for each final style attribute.
}
\description{
Merges the default node styles (`NODE_STYLES`) with any user-defined
  `style_overrides` from the recipe, and then joins these final styles to
  the nodes table based on `node_type`.
}
\keyword{internal}
