% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation_helper.R
\name{.assert_argument_is_plausible_string_selector}
\alias{.assert_argument_is_plausible_string_selector}
\title{Assert that a Quoted Argument for String Selection is Plausible}
\usage{
.assert_argument_is_plausible_string_selector(
  quo_string_selection,
  allowed_strings,
  arg_name_for_msg,
  allow_null = TRUE
)
}
\value{
Invisibly returns `NULL` if validation passes, otherwise stops with an error.
}
\description{
Performs basic checks on a quosure intended for tidyselect-style string
selection.
}
\keyword{internal}
