% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-configuration.R
\name{.build_model_hierarchy}
\alias{.build_model_hierarchy}
\title{Build Model Hierarchy Component}
\usage{
.build_model_hierarchy(features, user_args)
}
\arguments{
\item{features}{A list of detected model features.}

\item{user_args}{A list of user arguments.}
}
\value{
A list containing only the `$split_hierarchy`.
}
\description{
Determines the split hierarchy based on model features and user arguments.
}
\keyword{internal}
