% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-preparation.R
\name{.determine_final_node_labels}
\alias{.determine_final_node_labels}
\title{Determine Final Node Labels}
\usage{
.determine_final_node_labels(nodes_data, user_labels, value_col_name)
}
\arguments{
\item{nodes_data}{The input node data table, which must contain a 'label' column.}

\item{user_labels}{A named list of user-defined labels.}

\item{value_col_name}{The name of the column containing the numeric estimate.}
}
\value{
The `nodes_data` table with an updated `label` column.
}
\description{
Determines the final display label for each node. It takes the
  base label from the analysis phase, applies any user-defined overrides,
  and then sets the final label for special node types like variances and
  intercepts.
}
\keyword{internal}
