% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{.new_lavaan_layout}
\alias{.new_lavaan_layout}
\title{Internal S3 constructor for a `lavaan_layout` object.}
\usage{
.new_lavaan_layout(...)
}
\arguments{
\item{...}{A named list of layout results.}
}
\value{
An object of class `lavaan_layout`.
}
\description{
Creates the container for the results of the `layout` phase.
}
\keyword{internal}
