% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{.new_lavaan_model_structure}
\alias{.new_lavaan_model_structure}
\title{Internal S3 constructor for a `lavaan_model_structure` object.}
\usage{
.new_lavaan_model_structure(...)
}
\arguments{
\item{...}{A named list of analysis results (e.g., `features`, `layout`).}
}
\value{
An object of class `lavaan_model_structure`.
}
\description{
Creates the container for the results of the `analyze` phase.
}
\keyword{internal}
