% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{.new_lavaan_parameter_table}
\alias{.new_lavaan_parameter_table}
\title{Internal S3 constructor for a `lavaan_parameter_table` object.}
\usage{
.new_lavaan_parameter_table(param_table)
}
\arguments{
\item{...}{A named list, expected to contain the `param_table`.}
}
\value{
An object of class `lavaan_parameter_table`.
}
\description{
A simple wrapper to assign a class to the parameter table,
  enabling S3 dispatch.
}
\keyword{internal}
