% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-configuration.R
\name{.validate_user_args}
\alias{.validate_user_args}
\title{Validate User Arguments}
\usage{
.validate_user_args(user_args, data_context)
}
\arguments{
\item{user_args}{A list of normalized and defaulted user arguments.}

\item{data_context}{The `param_table` for context-dependent checks.}
}
\value{
The `user_args` list, passed through invisibly if validation succeeds.
}
\description{
A wrapper that calls the main validation engine.
}
\keyword{internal}
