% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavaanReportR-package.R
\docType{package}
\name{lavaanReportR-package}
\alias{lavaanReportR}
\alias{lavaanReportR-package}
\title{lavaanReportR: Streamlined Reporting for Multilevel SEM Models from 'lavaan'}
\description{
Provides auxiliary functions for the automatic generation of publication-ready formatted tables of coefficients, fit indices, and model comparisons for multilevel structural equation models (SEM) fitted with the 'lavaan' package. It aims to simplify the reporting workflow by producing tables suitable for direct inclusion in Quarto documents, with a focus on handling complex multilevel data structures (multiple levels, clusters, and groups) correctly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/orehren/lavaanReportR}
  \item Report bugs at \url{https://github.com/orehren/lavaanReportR/issues}
}

}
\author{
\strong{Maintainer}: Oliver Rehren \email{o.rehren@gmail.com} (\href{https://orcid.org/0000-0001-7602-8219}{ORCID})

}
\keyword{internal}
