% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02a-layout.R
\name{layout.lavaan_plot_config}
\alias{layout.lavaan_plot_config}
\title{Calculate Layout for a Plot}
\usage{
\method{layout}{lavaan_plot_config}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lavaan_plot_config}.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{lavaan_layout} object containing the calculated layout information.
}
\description{
This is the third phase of the `lavaanReportR` workflow. The
  `layout` function takes a `lavaan_plot_config` object and calculates the
  visual layout of the graph.
}
