% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{layout}
\alias{layout}
\title{Calculate Layout for a Plot}
\usage{
layout(x, ...)
}
\arguments{
\item{x}{The configured object to process.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A "layout" object, containing the calculated layout information.
}
\description{
A generic S3 function that dispatches to a specific method to
  calculate the visual layout of the graph based on the model structure and
  configuration.
}
\keyword{internal}
