% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{render}
\alias{render}
\title{Render a Built Graph Object}
\usage{
render(x, ...)
}
\arguments{
\item{x}{The built object to render.}

\item{...}{Additional arguments passed to methods.}
}
\value{
A rendered object, typically from `DiagrammeR` or `ggplot2`.
}
\description{
A generic S3 function that dispatches to a specific method to
  take a built object (like DOT code) and render it into a visual output.
}
\keyword{internal}
