% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{analyze}
\alias{analyze}
\title{Analyze an Input Object}
\usage{
analyze(x, ...)
}
\arguments{
\item{x}{The raw object to analyze.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An "analyzed" object, containing the results of the analysis.
}
\description{
A generic S3 function that dispatches to a specific method to
  perform analysis on a raw input object.
}
\keyword{internal}
