% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{layout}
\alias{layout}
\title{Calculate Node Layout Coordinates}
\usage{
layout(object, ...)
}
\arguments{
\item{object}{An object representing the graph to be laid out.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A new object with layout information added.
}
\description{
An S3 generic for calculating the x/y coordinates for graph nodes.
}
\keyword{internal}
