% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-preparation.R
\name{prepare.lavaan_layout}
\alias{prepare.lavaan_layout}
\title{Prepare Final Data for Plotting}
\usage{
\method{prepare}{lavaan_layout}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lavaan_layout}.}

\item{...}{Not used.}
}
\value{
A \code{lavaan_graph} object containing the final, build-ready
  data.tables for nodes and edges, along with the final recipe.
}
\description{
This is the fourth phase of the `lavaanReportR` workflow. The
  `prepare` function takes the `lavaan_layout` object and applies all
  styling, labeling, and filtering rules to produce the final, "build-ready"
  data tables for nodes and edges.
}
