% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05-render-api.R
\name{render}
\alias{render}
\alias{render.lavaan_dot_code}
\title{Render the Final Plot}
\usage{
\method{render}{lavaan_dot_code}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lavaan_dot_code}.}

\item{...}{Not used.}
}
\value{
A \code{dgr_graph} object from the `DiagrammeR` package.
}
\description{
This is the final phase of the `lavaanReportR` workflow. The
`render` function takes the `lavaan_dot_code` object and uses the `DiagrammeR`
package to render the final graph.
}
\details{
This is the simplest phase of the workflow. It acts as a lightweight wrapper
around `DiagrammeR::grViz()`, passing the generated DOT code to be rendered
into a final plot object.
}
\examples{
\dontrun{
  # Assuming 'dot_code' is from the build() step
  final_plot <- render(dot_code)
  # The plot can now be displayed
  final_plot
}
}
\seealso{
\code{\link{analyze}}, \code{\link{configure_plot}}, \code{\link{prepare}}, \code{\link{build}}
}
